+++ cs-lib/lubx/Camera3d.cs
@@ -2,7 +2,7 @@
 // opts は options class (Camera3dOpts) で受ける (stub の PassOpts / DrawOpts
 // と同じ流儀)。
 // optional は nullable フィールド + ?? で受ける。Gfx.size の multi-return は
-// out 引数で受け、aspect の除算は (double) cast で整数除算を避ける。
+// out 引数で受け、aspect の除算は (float) cast で整数除算を避ける。
 
 /// <summary>Camera3d.vp のオプション。eye / target は必須、他は省略可。</summary>
 using static Lub;
@@ -12,10 +12,10 @@ public class Camera3dOpts
     public Vec3 Eye = new Vec3(0, 0, 0);
     public Vec3 Target = new Vec3(0, 0, 0);
     public Vec3? Up;
-    public double? Fov;
-    public double? Near;
-    public double? Far;
-    public double? Aspect;
+    public float? Fov;
+    public float? Near;
+    public float? Far;
+    public float? Aspect;
 }
 
 /// <summary>3D カメラ定型。perspective + lookAt から view-projection を
@@ -27,18 +27,18 @@ public static class Camera3d
     public static Mat4 Vp(Camera3dOpts opts)
     {
         var up = opts.Up ?? new Vec3(0, 1, 0);
-        var fov = opts.Fov ?? 60.0;
-        var near = opts.Near ?? 0.1;
-        var far = opts.Far ?? 100.0;
-        double aspect;
+        var fov = opts.Fov ?? 60.0f;
+        var near = opts.Near ?? 0.1f;
+        var far = opts.Far ?? 100.0f;
+        float aspect;
         if (opts.Aspect != null)
         {
-            aspect = opts.Aspect ?? 1.0;
+            aspect = opts.Aspect ?? 1.0f;
         }
         else
         {
             Gfx.Size(out var gw, out var gh);
-            aspect = (double)gw / gh;
+            aspect = (float)gw / gh;
         }
         var proj = Mat4.PerspectiveLh(fov, aspect, near, far);
         var view = Mat4.LookAtLh(opts.Eye, opts.Target, up);
