+++ cs-lib/lubx/Sdf.cs
@@ -14,26 +14,26 @@ using static Lub;
 public class SdfNode
 {
     public Lub.Mesh.SdfOp Op;
-    public List<double> Params;
+    public List<float> Params;
     public string? Name;
     public SdfNode? C;
     public SdfNode? A;
     public SdfNode? B;
 
-    public SdfNode(Lub.Mesh.SdfOp op, List<double> parameters)
+    public SdfNode(Lub.Mesh.SdfOp op, List<float> parameters)
     {
         this.Op = op;
         this.Params = parameters;
     }
 
-    static SdfNode Unary(Lub.Mesh.SdfOp op, List<double> parameters, SdfNode c)
+    static SdfNode Unary(Lub.Mesh.SdfOp op, List<float> parameters, SdfNode c)
     {
         var n = new SdfNode(op, parameters);
         n.C = c;
         return n;
     }
 
-    static SdfNode Binary(Lub.Mesh.SdfOp op, List<double> parameters, SdfNode a,
+    static SdfNode Binary(Lub.Mesh.SdfOp op, List<float> parameters, SdfNode a,
         SdfNode b)
     {
         var n = new SdfNode(op, parameters);
@@ -42,57 +42,57 @@ public class SdfNode
         return n;
     }
 
-    public SdfNode Move(double x, double y, double z) =>
-        Unary(Lub.Mesh.SdfOp.Move, new List<double> { x, y, z }, this);
+    public SdfNode Move(float x, float y, float z) =>
+        Unary(Lub.Mesh.SdfOp.Move, new List<float> { x, y, z }, this);
 
     /// <summary>`axis` 回りに `rad` ラジアン回す。</summary>
-    public SdfNode Rotate(Vec3 axis, double rad)
+    public SdfNode Rotate(Vec3 axis, float rad)
     {
         var q = Quat.FromAxisAngle(axis, rad);
-        return Unary(Lub.Mesh.SdfOp.Rotate, new List<double> { q.X, q.Y, q.Z, q.W },
+        return Unary(Lub.Mesh.SdfOp.Rotate, new List<float> { q.X, q.Y, q.Z, q.W },
             this);
     }
 
     /// <summary>uniform スケール (`s` &gt; 0)。</summary>
-    public SdfNode Scale(double s) =>
-        Unary(Lub.Mesh.SdfOp.Scale, new List<double> { s }, this);
+    public SdfNode Scale(float s) =>
+        Unary(Lub.Mesh.SdfOp.Scale, new List<float> { s }, this);
 
     /// <summary>X 対称 (|x| 折り畳み)。</summary>
     public SdfNode MirrorX() =>
-        Unary(Lub.Mesh.SdfOp.MirrorX, new List<double>(), this);
+        Unary(Lub.Mesh.SdfOp.MirrorX, new List<float>(), this);
 
     /// <summary>サブツリーに材質 (albedo + metallic/roughness) を与える。
     /// innermost の paint が勝つ。`rgb` は 0xRRGGBB (metallic 省略で 0.0、
     /// roughness 省略で 0.8)。smin/ssub では距離と同じ blend で材質も混ざり、
     /// subtract/ssub の切断面には cutter の材質が出る。</summary>
-    public SdfNode Paint(int rgb, double? metallic = null,
-        double? roughness = null) =>
-        Unary(Lub.Mesh.SdfOp.Paint, new List<double>
+    public SdfNode Paint(int rgb, float? metallic = null,
+        float? roughness = null) =>
+        Unary(Lub.Mesh.SdfOp.Paint, new List<float>
         {
-            Math.Floor(rgb / 65536.0) % 256 / 255.0,
-            Math.Floor(rgb / 256.0) % 256 / 255.0,
-            rgb % 256 / 255.0,
-            metallic ?? 0.0,
-            roughness ?? 0.8,
+            (float)Math.Floor(rgb / 65536.0f) % 256 / 255.0f,
+            (float)Math.Floor(rgb / 256.0f) % 256 / 255.0f,
+            rgb % 256 / 255.0f,
+            metallic ?? 0.0f,
+            roughness ?? 0.8f,
         }, this);
 
     public SdfNode Union(SdfNode b) =>
-        Binary(Lub.Mesh.SdfOp.Union, new List<double>(), this, b);
+        Binary(Lub.Mesh.SdfOp.Union, new List<float>(), this, b);
 
     /// <summary>smooth union。`k` が blend 幅。</summary>
-    public SdfNode Smin(SdfNode b, double k) =>
-        Binary(Lub.Mesh.SdfOp.Smin, new List<double> { k }, this, b);
+    public SdfNode Smin(SdfNode b, float k) =>
+        Binary(Lub.Mesh.SdfOp.Smin, new List<float> { k }, this, b);
 
     /// <summary>`b` をくり抜く。</summary>
     public SdfNode Subtract(SdfNode b) =>
-        Binary(Lub.Mesh.SdfOp.Subtract, new List<double>(), this, b);
+        Binary(Lub.Mesh.SdfOp.Subtract, new List<float>(), this, b);
 
     /// <summary>smooth subtraction。`k` が縁の丸まり幅。</summary>
-    public SdfNode Ssub(SdfNode b, double k) =>
-        Binary(Lub.Mesh.SdfOp.Ssub, new List<double> { k }, this, b);
+    public SdfNode Ssub(SdfNode b, float k) =>
+        Binary(Lub.Mesh.SdfOp.Ssub, new List<float> { k }, this, b);
 
     public SdfNode Intersect(SdfNode b) =>
-        Binary(Lub.Mesh.SdfOp.Intersect, new List<double>(), this, b);
+        Binary(Lub.Mesh.SdfOp.Intersect, new List<float>(), this, b);
 
     /// <summary>サブツリーを skinning 部位として宣言する。`pivot` は関節位置
     /// (model 空間)。メッシュ化時に頂点ごとの部位距離から重みが焼かれる。
@@ -101,7 +101,7 @@ public class SdfNode
     public SdfNode Bone(string name, Vec3 pivot)
     {
         var n = Unary(Lub.Mesh.SdfOp.Bone,
-            new List<double> { pivot.X, pivot.Y, pivot.Z }, this);
+            new List<float> { pivot.X, pivot.Y, pivot.Z }, this);
         n.Name = name;
         return n;
     }
@@ -116,21 +116,21 @@ public class SdfNode
 /// </code></summary>
 public static class Sdf
 {
-    public static SdfNode Sphere(double r) =>
-        new SdfNode(Lub.Mesh.SdfOp.Sphere, new List<double> { r });
+    public static SdfNode Sphere(float r) =>
+        new SdfNode(Lub.Mesh.SdfOp.Sphere, new List<float> { r });
 
     /// <summary>half extents の直方体。</summary>
-    public static SdfNode Box(double hx, double hy, double hz) =>
-        new SdfNode(Lub.Mesh.SdfOp.Box, new List<double> { hx, hy, hz });
+    public static SdfNode Box(float hx, float hy, float hz) =>
+        new SdfNode(Lub.Mesh.SdfOp.Box, new List<float> { hx, hy, hz });
 
     /// <summary>線分 `a`-`b` を軸とする半径 `r` のカプセル。</summary>
-    public static SdfNode Capsule(Vec3 a, Vec3 b, double r) =>
+    public static SdfNode Capsule(Vec3 a, Vec3 b, float r) =>
         new SdfNode(Lub.Mesh.SdfOp.Capsule,
-            new List<double> { a.X, a.Y, a.Z, b.X, b.Y, b.Z, r });
+            new List<float> { a.X, a.Y, a.Z, b.X, b.Y, b.Z, r });
 
     /// <summary>XZ 平面に寝たトーラス。</summary>
-    public static SdfNode Torus(double rMajor, double rMinor) =>
-        new SdfNode(Lub.Mesh.SdfOp.Torus, new List<double> { rMajor, rMinor });
+    public static SdfNode Torus(float rMajor, float rMinor) =>
+        new SdfNode(Lub.Mesh.SdfOp.Torus, new List<float> { rMajor, rMinor });
 
     /// <summary>op の表示名 (SdfPanel のラベル用)。</summary>
     public static string OpName(Lub.Mesh.SdfOp op)
@@ -177,7 +177,7 @@ public static class Sdf
     /// <summary>ツリーをメッシュ化する。`n` は最長軸の cell 数 (bounds は
     /// 自動)。bone ノードがあれば skinning 情報も焼かれる
     /// (`skinK` = 重みの blend 幅)。</summary>
-    public static MeshData Mesh(SdfNode root, int n, double? skinK = null)
+    public static MeshData Mesh(SdfNode root, int n, float? skinK = null)
     {
         var nodes = new List<SdfNodeDesc>();
         var r = Flatten(root, nodes);
