+++ cs-lib/lubx/Sfx.cs
@@ -1,5 +1,5 @@
 // 実装ライブラリ lubx の Sfx。
-// 波形は List<double>.Add で積む (TinyC# の List は最初から Lua array table)。
+// 波形は List<float>.Add で積む (TinyC# の List は最初から Lua array table)。
 // 切り捨ては値が非負なので Math.Floor、i % 16 == 0 は整数剰余を避けて
 // (i & 15) == 0。
 using System.Collections.Generic;
@@ -16,10 +16,10 @@ public static class Sfx
 {
     public const int Rate = 44100;
 
-    private static Dictionary<string, List<double>> cache = new Dictionary<string, List<double>>();
+    private static Dictionary<string, List<float>> cache = new Dictionary<string, List<float>>();
 
     /// <summary>矩形波 blip。freq0→freq1 へスイープしつつ指数減衰 (exp(-5u))。snd handle を返す。</summary>
-    public static int Blip(double freq0, double freq1, double dur, double vol)
+    public static int Blip(float freq0, float freq1, float dur, float vol)
     {
         var key = "blip:" + freq0 + ":" + freq1 + ":" + dur + ":" + vol;
         if (cache.TryGetValue(key, out var cached))
@@ -27,22 +27,22 @@ public static class Sfx
             return Audio.Snd(key, cached, 1, Rate, 1);
         }
         var n = (int)System.Math.Floor(dur * Rate);
-        var samples = new List<double>();
-        var phase = 0.0;
+        var samples = new List<float>();
+        var phase = 0.0f;
         for (var i = 0; i < n; i++)
         {
-            var u = (double)i / n;
+            var u = (float)i / n;
             var freq = freq0 + (freq1 - freq0) * u;
             phase += freq / Rate;
-            var env = System.Math.Exp(-5.0 * u);
-            samples.Add((phase % 1.0 < 0.5 ? 1.0 : -1.0) * env * vol);
+            var env = (float)System.Math.Exp(-5.0f * u);
+            samples.Add((phase % 1.0f < 0.5f ? 1.0f : -1.0f) * env * vol);
         }
         cache[key] = samples;
         return Audio.Snd(key, samples, 1, Rate, 1);
     }
 
     /// <summary>ノイズバースト。指数減衰 (exp(-4u))、16 sample ごとにホールド更新。snd handle を返す。</summary>
-    public static int Noise(double dur, double vol, int? seed = null)
+    public static int Noise(float dur, float vol, int? seed = null)
     {
         var s = seed ?? 0x12345678;
         var key = "noise:" + dur + ":" + vol + ":" + s;
@@ -51,17 +51,17 @@ public static class Sfx
             return Audio.Snd(key, cached, 1, Rate, 1);
         }
         var n = (int)System.Math.Floor(dur * Rate);
-        var samples = new List<double>();
+        var samples = new List<float>();
         var r = new Rand(s);
-        var hold = 0.0;
+        var hold = 0.0f;
         for (var i = 0; i < n; i++)
         {
             if ((i & 15) == 0)
             {
-                hold = r.NextFloat() * 2.0 - 1.0;
+                hold = r.NextFloat() * 2.0f - 1.0f;
             }
-            var u = (double)i / n;
-            samples.Add(hold * System.Math.Exp(-4.0 * u) * vol);
+            var u = (float)i / n;
+            samples.Add(hold * (float)System.Math.Exp(-4.0f * u) * vol);
         }
         cache[key] = samples;
         return Audio.Snd(key, samples, 1, Rate, 1);
